+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [0.10.0] 2025-03-20
+
+### Added
+
+- `extract coverage-xml` command:
+  - Extract metadata from a Coverage XML output files.
+
 ## [0.9.1] 2025-03-19
 
 ### Fixed